# openapi.legacy.yaml
openapi: 3.0.3
info:
  title: Tierfog
  version: 0.1.0
  description: Multi-tier edge/fog fabric converging IoT protocols for in-place federated analytics.
  x-product-slug: ai-decentralized-tierfog
  x-source: ai-in-decentralized+ai/muraliiot-edge-ai-networks-180601042020/
servers:
- url: https://api.example.com
...
paths:
...
  /v1/tiers:
    get:
      operationId: listTier
      tags:
      - Tiers
      security:
      - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelopeTierNode'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - $ref: '#/components/parameters/Cursor'
    post:
      operationId: registerTier
      tags:
      - Tiers
      security:
      - BearerAuth: []
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeTierNode'
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TierNodeCreate'
  /v1/domains:
    get:
      operationId: listProtocolDomains
      tags:
      - Domains
      security:
      - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelopeProtocolDomain'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - $ref: '#/components/parameters/Cursor'
    post:
      operationId: onboardProtocolDomain
      tags:
      - Domains
      security:
      - BearerAuth: []
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeProtocolDomain'
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtocolDomainCreate'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    ResponseMeta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    TierNode:
      type: object
      required:
      - id
      - siteId
      - role
      properties:
        id:
          type: string
        siteId:
          type: string
        role:
          type: string
          enum:
          - device
          - fog
          - edge
          - cloud
        parentTierId:
          type: string
        status:
          type: string
          enum:
          - active
          - quarantined
    TierNodeCreate:
      type: object
      required:
      - siteId
      - role
      properties:
        siteId:
          type: string
        role:
          type: string
          enum:
          - device
          - fog
          - edge
          - cloud
        parentTierId:
          type: string
    ProtocolDomain:
      type: object
      required:
      - id
      - protocolFamily
      - tierId
      properties:
        id:
          type: string
        protocolFamily:
          type: string
          enum:
          - mqtt_coap
          - rpl_6lowpan
          - bt_mesh
          - v2x
          - http_rich
          - other
        tierId:
          type: string
        gatewayConfigVersion:
          type: string
    ProtocolDomainCreate:
      type: object
      required:
      - protocolFamily
      - tierId
      properties:
        protocolFamily:
          type: string
        tierId:
          type: string
        gatewayConfigVersion:
          type: string
...
    DataEnvelopeTierNode:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/TierNode'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ListEnvelopeTierNode:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          required:
          - items
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/TierNode'
            nextCursor:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    DataEnvelopeProtocolDomain:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/ProtocolDomain'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ListEnvelopeProtocolDomain:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          required:
          - items
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ProtocolDomain'
            nextCursor:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...