# platform/api-server/src/domains/protocol/handlers/domain.handlers.ts
/**
 * Handlers for Domain
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: protocol
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ProtocolDomainModule } from "../dependencies/protocol-ddd.dependencies.js";
import type { GetProtocolDomainInput, ListProtocolDomainsInput, OnboardProtocolDomainInput } from "@tierfog/services/protocol";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/domains
 * Operation: listProtocolDomains
 */
export async function listProtocolDomains(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], tierId?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ProtocolDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListProtocolDomainsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.domains.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/domains
 * Operation: onboardProtocolDomain
 */
export async function onboardProtocolDomain(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ProtocolDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: OnboardProtocolDomainInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.domains.get.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/domains/{domainId}
 * Operation: getProtocolDomain
 */
export async function getProtocolDomain(
  request: FastifyRequest<{ Params: { domainId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: ProtocolDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetProtocolDomainInput = {
    domainId: request.params.domainId,
  };
  const result = await deps.useCases.domains.get.execute(input);
  return reply.code(200).send(result);
}