# platform/api-server/src/domains/tiers/handlers/tier.handlers.ts
/**
 * Handlers for Tier
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: tiers
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { TiersDomainModule } from "../dependencies/tiers-ddd.dependencies.js";
import type { GetTierInput, ListTiersInput, RegisterTierInput } from "@tierfog/services/tiers";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/tiers
 * Operation: listTiers
 */
export async function listTiers(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], siteId?: string | string[], status?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: TiersDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListTiersInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.tiers.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/tiers
 * Operation: registerTier
 */
export async function registerTier(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: TiersDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RegisterTierInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.tiers.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/tiers/{tierId}
 * Operation: getTier
 */
export async function getTier(
  request: FastifyRequest<{ Params: { tierId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: TiersDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetTierInput = {
    tierId: request.params.tierId,
  };
  const result = await deps.useCases.tiers.get.execute(input);
  return reply.code(200).send(result);
}